Added namespace to Auth tests
